+++ betamax/__init__.py
@@ -1,5 +1,6 @@
 """
-betamax
+betamax.
+
 =======
 
 See http://betamax.rtfd.org/ for documentation.
@@ -9,15 +10,17 @@ See http://betamax.rtfd.org/ for documentation.
 
 """
 
+from .decorator import use_cassette
 from .exceptions import BetamaxError
-from .recorder import Betamax
 from .matchers import BaseMatcher
+from .recorder import Betamax
 from .serializers import BaseSerializer
 
-__all__ = ('BetamaxError', 'Betamax', 'BaseMatcher', 'BaseSerializer')
+__all__ = ('BetamaxError', 'Betamax', 'BaseMatcher', 'BaseSerializer',
+           'use_cassette')
 __author__ = 'Ian Cordasco'
-__copyright__ = 'Copyright 2013 Ian Cordasco'
+__copyright__ = 'Copyright 2013-2014 Ian Cordasco'
 __license__ = 'Apache 2.0'
 __title__ = 'betamax'
-__version__ = '0.4.0'
+__version__ = '0.5.1'
 __version_info__ = tuple(int(i) for i in __version__.split('.'))