+++ betamax/adapter.py
@@ -1,3 +1,10 @@
+"""
+betamax.adapter.
+
+==============
+
+adapter for betamax
+"""
 import os
 
 from .cassette import Cassette
@@ -7,7 +14,6 @@ from requests.adapters import BaseAdapter, HTTPAdapter
 
 
 class BetamaxAdapter(BaseAdapter):
-
     """This object is an implementation detail of the library.
 
     It is not meant to be a public API and is not exported as such.
@@ -24,19 +30,33 @@ class BetamaxAdapter(BaseAdapter):
         self.options = {}
 
     def cassette_exists(self):
+        """Check if cassette exists on file system.
+
+        :returns: bool -- True if exists, False otherwise
+        """
         if self.cassette_name and os.path.exists(self.cassette_name):
             return True
         return False
 
     def close(self):
+        """Propagate close to underlying adapter."""
         self.http_adapter.close()
 
     def eject_cassette(self):
+        """Eject currently loaded cassette."""
         if self.cassette:
             self.cassette.eject()
         self.cassette = None  # Allow self.cassette to be garbage-collected
 
     def load_cassette(self, cassette_name, serialize, options):
+        """Load cassette.
+
+        Loads a previously serialized http response as a cassette
+
+        :param str cassette_name: (required), name of cassette
+        :param str serialize: (required), type of serialization i.e 'json'
+        :options dict options: (required), options for cassette
+        """
         self.cassette_name = cassette_name
         self.serialize = serialize
         self.options.update(options.items())
@@ -73,6 +93,11 @@ class BetamaxAdapter(BaseAdapter):
 
     def send(self, request, stream=False, timeout=None, verify=True,
              cert=None, proxies=None):
+        """Send request.
+
+        :param request request: request
+        :returns: A Response object
+        """
         interaction = None
 
         if not self.cassette:
@@ -96,6 +121,21 @@ class BetamaxAdapter(BaseAdapter):
 
     def send_and_record(self, request, stream=False, timeout=None,
                         verify=True, cert=None, proxies=None):
+        """Send request and record response.
+
+        The response will be serialized and saved to a
+        cassette which can be replayed in the future.
+
+        :param request request: request
+        :param bool stream: (optional) defer download until content is accessed
+        :param float timeout: (optional) time to wait for a response
+        :param bool verify: (optional) verify SSL certificate
+        :param str cert: (optional) path to SSL client
+        :param proxies dict: (optional) mapping protocol to URL of the proxy
+        :return: Iteraction
+        :rtype: class:`betamax.cassette.iteraction`
+
+        """
         adapter = self.find_adapter(request.url)
         response = adapter.send(
             request, stream=True, timeout=timeout, verify=verify,
@@ -105,6 +145,13 @@ class BetamaxAdapter(BaseAdapter):
         return self.cassette.interactions[-1]
 
     def find_adapter(self, url):
+        """Find adapter.
+
+        Searches for an existing adapter where the url and prefix match.
+
+        :param url str: (required) url of the adapter
+        :returns: betamax adapter
+        """
         for (prefix, adapter) in self.old_adapters.items():
 
             if url.lower().startswith(prefix):
@@ -125,6 +172,7 @@ The settings on the cassette are:
 
 
 def unhandled_request_message(request, cassette):
+    """Generate exception for unhandled requests."""
     return UNHANDLED_REQUEST_EXCEPTION.format(
         url=request.url, cassette_file_path=cassette.cassette_name,
         cassette_record_mode=cassette.record_mode,