+++ betamax/cassette/__init__.py
@@ -1,5 +1,4 @@
 from .cassette import Cassette
 from .interaction import Interaction
-from .mock_response import MockHTTPResponse
 
-__all__ = ['Cassette', 'Interaction', 'MockHTTPResponse']
+__all__ = ('Cassette', 'Interaction')