+++ betamax/matchers/__init__.py
@@ -1,5 +1,3 @@
-matcher_registry = {}
-
 from .base import BaseMatcher
 from .body import BodyMatcher
 from .digest_auth import DigestAuthMatcher
@@ -10,10 +8,11 @@ from .path import PathMatcher
 from .query import QueryMatcher
 from .uri import URIMatcher
 
+matcher_registry = {}
 
-__all__ = ['BaseMatcher', 'BodyMatcher', 'DigestAuthMatcher',
+__all__ = ('BaseMatcher', 'BodyMatcher', 'DigestAuthMatcher',
            'HeadersMatcher', 'HostMatcher', 'MethodMatcher', 'PathMatcher',
-           'QueryMatcher', 'URIMatcher']
+           'QueryMatcher', 'URIMatcher', 'matcher_registry')
 
 
 _matchers = [BodyMatcher, DigestAuthMatcher, HeadersMatcher, HostMatcher,