+++ tests/integration/test_record_modes.py
@@ -24,9 +24,16 @@ class TestRecordOnce(IntegrationHelper):
             assert betamax.current_cassette.interactions != []
             assert len(betamax.current_cassette.interactions) == 1
             r1 = s.get('http://httpbin.org/get')
-            assert len(betamax.current_cassette.interactions) == 1
+            assert len(betamax.current_cassette.interactions) == 2
             assert r1.status_code == 200
-            assert r0.headers == r1.headers
+            r0_headers = r0.headers.copy()
+            r0_headers.pop('Date')
+            r1_headers = r1.headers.copy()
+            r1_headers.pop('Date')
+            # NOTE(sigmavirus24): This fails if the second request is
+            # technically a second later. Ignoring the Date headers allows
+            # this test to succeed.
+            assert r0_headers == r1_headers
             assert r0.content == r1.content
 
 
@@ -61,14 +68,14 @@ class TestRecordNewEpisodes(IntegrationHelper):
             cassette = betamax.current_cassette
             self.cassette_path = cassette.cassette_path
             assert cassette.interactions != []
-            assert len(cassette.interactions) == 3
+            assert len(cassette.interactions) == 4
             assert cassette.is_empty() is False
             s.get('https://httpbin.org/get')
-            assert len(cassette.interactions) == 4
+            assert len(cassette.interactions) == 5
 
         with Betamax(s).use_cassette('test_record_new') as betamax:
             cassette = betamax.current_cassette
-            assert len(cassette.interactions) == 4
+            assert len(cassette.interactions) == 5
             r = s.get('https://httpbin.org/get')
             assert r.status_code == 200
 
@@ -89,6 +96,7 @@ class TestRecordAll(IntegrationHelper):
         with Betamax(self.session).use_cassette('test_record_all'):
             self.session.get('http://httpbin.org/get')
             self.session.get('http://httpbin.org/redirect/2')
+            self.session.get('http://httpbin.org/get')
 
     def test_records_new_interactions(self):
         s = self.session
@@ -97,13 +105,13 @@ class TestRecordAll(IntegrationHelper):
             cassette = betamax.current_cassette
             self.cassette_path = cassette.cassette_path
             assert cassette.interactions != []
-            assert len(cassette.interactions) == 3
+            assert len(cassette.interactions) == 5
             assert cassette.is_empty() is False
             s.post('http://httpbin.org/post', data={'foo': 'bar'})
-            assert len(cassette.interactions) == 4
+            assert len(cassette.interactions) == 6
 
         with Betamax(s).use_cassette('test_record_all') as betamax:
-            assert len(betamax.current_cassette.interactions) == 4
+            assert len(betamax.current_cassette.interactions) == 6
 
     def test_replaces_old_interactions(self):
         s = self.session
@@ -112,7 +120,7 @@ class TestRecordAll(IntegrationHelper):
             cassette = betamax.current_cassette
             self.cassette_path = cassette.cassette_path
             assert cassette.interactions != []
-            assert len(cassette.interactions) == 3
+            assert len(cassette.interactions) == 5
             assert cassette.is_empty() is False
             s.get('http://httpbin.org/get')
-            assert len(cassette.interactions) == 3
+            assert len(cassette.interactions) == 5