+++ tests/regression/test_cassettes_retain_global_configuration.py
@@ -8,13 +8,13 @@ from requests import Session
 class TestCassetteRecordMode(unittest.TestCase):
     def setUp(self):
         with Betamax.configure() as config:
-            config.default_cassette_options['record_mode'] = 'never'
+            config.default_cassette_options['record_mode'] = 'none'
 
     def tearDown(self):
         with Betamax.configure() as config:
             config.default_cassette_options['record_mode'] = 'once'
 
-    def test_record_mode_is_never(self):
+    def test_record_mode_is_none(self):
         s = Session()
         with pytest.raises(ValueError):
             with Betamax(s) as recorder:
@@ -23,4 +23,4 @@ class TestCassetteRecordMode(unittest.TestCase):
 
     def test_class_variables_retain_their_value(self):
         opts = cassette.Cassette.default_cassette_options
-        assert opts['record_mode'] == 'never'
+        assert opts['record_mode'] == 'none'